% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.groupedData.R
\name{autoplot.groupedData}
\alias{autoplot.groupedData}
\title{Autoplot objects from \pkg{datasets}}
\usage{
\method{autoplot}{groupedData}(object, ...)
}
\arguments{
\item{object}{..}

\item{...}{..}
}
\description{
..
}
\examples{
autoplot(Indometh)
autoplot(Loblolly)
autoplot(ChickWeight) # Chick and Diet already factor

autoplot(nlme::Soybean) # two outer variables
autoplot(nlme::Glucose2) # nested groups

}
