% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{autoplot.lme}
\alias{autoplot.lme}
\title{autoplot objects from \pkg{nlme}}
\usage{
\method{autoplot}{lme}(object, level = 0L, ...)
}
\arguments{
\item{object}{..}

\item{level}{..}

\item{...}{..}
}
\description{
..
}
\examples{
# not written yet

}
