% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{coef_.lme}
\alias{coef_.lme}
\title{Fixed Effect of \link[nlme]{lme} Object}
\usage{
coef_.lme(x)
}
\arguments{
\item{x}{\link[nlme]{lme} or \link[nlme]{gls} object}
}
\description{
Fixed Effect of \link[nlme]{lme} Object
}
\note{
We already have \code{nlme:::coef.lme}

Function \code{\link[=coef_.lme]{coef_.lme()}} is actually \code{nlme:::fixef.lme}.
}
