% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint2.R
\name{confint2}
\alias{confint2}
\alias{confint_.lme}
\alias{confint_.gls}
\alias{confint_.intervals.lme}
\alias{confint_.intervals.gls}
\title{confint2}
\usage{
confint_.lme(x, level = 0.95, ...)

confint_.gls(x, level = 0.95, ...)

confint_.intervals.lme(x, ...)

confint_.intervals.gls(x, ...)
}
\arguments{
\item{x}{..}

\item{level}{..}

\item{...}{..}
}
\description{
confint2
}
\examples{
ortho1 |> confint_.lme()
ovary1 |> confint_.gls()
}
