% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_nlme}
\alias{desc_nlme}
\alias{desc_.lme}
\alias{desc_.gls}
\title{Model Description}
\usage{
\method{desc_}{lme}(x)

\method{desc_}{gls}(x)
}
\arguments{
\item{x}{\link[nlme]{lme} or \link[nlme]{gls} object}
}
\description{
Model Description
}
\keyword{internal}
