% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_nlme}
\alias{diagnose_nlme}
\alias{diagnose.lme}
\alias{diagnose.gls}
\title{Diagnose \link[nlme]{lme} and \link[nlme]{gls}}
\usage{
diagnose.lme(object, ...)

diagnose.gls(object, ...)
}
\arguments{
\item{object}{\link[nlme]{lme} or \link[nlme]{gls} object}

\item{...}{..}
}
\description{
..
}
\examples{
library(nlme)
lme(distance ~ age, data = Orthodont, keep.data = TRUE) |> 
 diagnose.lme(lty = 2, id = .05, adj = -.4)
lme(fixed = distance ~ Sex * I(age-11), 
 weights = varIdent(form = ~ 1 | Sex), data = Orthodont) |> diagnose.lme()
lme(Thickness ~ 1, data = Oxide) |> diagnose.lme()
# gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), 
#  data = Ovary, correlation = corAR1(form = ~ 1 | Mare)) |> diagnose.gls() # bug

}
