% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{lme_models}
\alias{lme_models}
\alias{ortho1}
\alias{ortho2}
\alias{oxide1}
\alias{ovary1}
\title{Pre-run models for Orthodont data}
\format{
..

An object of class \code{lme} of length 18.

An object of class \code{lme} of length 18.

An object of class \code{gls} of length 17.
}
\usage{
ortho1

ortho2

oxide1

ovary1
}
\description{
..
}
\keyword{datasets}
