% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_.R
\name{md_nlme}
\alias{md_nlme}
\alias{md_.lme}
\alias{md_.gls}
\title{R Markdown Lines for \code{lme} and \code{gls}}
\usage{
\method{md_}{lme}(x, xnm, ...)

\method{md_}{gls}(x, xnm, ...)
}
\arguments{
\item{x, xnm, ...}{..}
}
\description{
R Markdown Lines for \code{lme} and \code{gls}
}
\examples{
library(nlme)
m1 = lme(distance ~ age, data = Orthodont, keep.data = TRUE)
m2 = gls(follicles ~ sin(2*pi*Time) + cos(2*pi*Time), 
  data = Ovary, correlation = corAR1(form = ~ 1 | Mare)) 
  
library(ecip)
list('`lme`' = m1, '`gls`' = m2) |> fastmd::render_(file = 'lme_gls')
}
\keyword{internal}
