% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.frame.R
\name{model_frame_nlme}
\alias{model_frame_nlme}
\alias{model.frame.lme}
\alias{model.frame.gls}
\title{\link[stats]{model.frame} of \link[nlme]{gls} and \link[nlme]{lme} Objects}
\usage{
\method{model.frame}{lme}(formula, data = getData(formula), ...)

\method{model.frame}{gls}(formula, data = getData(formula), ...)
}
\arguments{
\item{formula}{\link[nlme]{gls} or \link[nlme]{lme} object}

\item{data}{..}

\item{...}{additional parameters of \link[stats]{model.frame.default}}
}
\description{
..
}
\note{
S3 generic \link[stats]{model.frame} is critical for function \link[stats]{.getXlevels}.

Without function \link{model.frame.lme},
S3 generic \link[stats]{model.frame}
dispatches to \link[stats]{model.frame.default} and
get \verb{$modelStruct} (as \link[nlme]{lme} object has no \verb{$model} element).
}
