% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobsText.R
\name{nobsText_nlme}
\alias{nobsText_nlme}
\alias{nobsText.lme}
\alias{nobsText.gls}
\title{nobsText}
\usage{
\method{nobsText}{lme}(x)

\method{nobsText}{gls}(x)
}
\arguments{
\item{x}{\link[nlme]{lme} or \link[nlme]{gls} object}
}
\description{
nobsText
}
\keyword{internal}
