% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{Sprintf.gls}
\alias{Sprintf.gls}
\title{Sprintf.gls}
\usage{
Sprintf.gls(model, ...)
}
\arguments{
\item{model}{\link[nlme]{lme} object}

\item{...}{..}
}
\description{
..
}
\examples{
Sprintf.gls(ovary1)

}
