% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{Sprintf.lme}
\alias{Sprintf.lme}
\title{Sprintf.lme}
\usage{
Sprintf.lme(model, ...)
}
\arguments{
\item{model}{\link[nlme]{lme} object}

\item{...}{..}
}
\description{
..
}
\examples{
Sprintf.lme(ortho1)

}
