% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{Sprintf_lme_gls}
\alias{Sprintf_lme_gls}
\alias{Sprintf.lme}
\alias{Sprintf.gls}
\title{Sprintf.lme}
\usage{
Sprintf.lme(x)

Sprintf.gls(x)
}
\arguments{
\item{x}{\link[nlme]{lme} object or \link[nlme]{gls} object}
}
\description{
..
}
\examples{
Sprintf.lme(ortho1)

Sprintf.gls(ovary1)
}
