% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint2.R
\name{confint2}
\alias{confint2}
\alias{confint2.lme}
\alias{confint2.gls}
\alias{confint2.intervals.lme}
\alias{confint2.intervals.gls}
\title{confint2}
\usage{
confint2.lme(x, level = 0.95, ...)

confint2.gls(x, level = 0.95, ...)

confint2.intervals.lme(x, ...)

confint2.intervals.gls(x, ...)
}
\arguments{
\item{x}{..}

\item{level}{..}

\item{...}{..}
}
\description{
confint2
}
\examples{
ortho1 |> confint2.lme()
ovary1 |> confint2.gls()
}
