% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint_nlme}
\alias{confint_nlme}
\alias{confint_.lme}
\alias{confint_.gls}
\alias{confint_.intervals.lme}
\alias{confint_.intervals.gls}
\title{confint_}
\usage{
confint_.lme(x, level = 0.95, ...)

confint_.gls(x, level = 0.95, ...)

confint_.intervals.lme(x, ...)

confint_.intervals.gls(x, ...)
}
\arguments{
\item{x}{..}

\item{level}{..}

\item{...}{..}
}
\description{
confint_
}
\examples{
ortho1 |> confint_.lme()
ovary1 |> confint_.gls()
}
