% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_nlme}
\alias{diagnose_nlme}
\alias{diagnose.lme}
\alias{diagnose.gls}
\title{Diagnose \link[nlme]{lme} and \link[nlme]{gls}}
\usage{
diagnose.lme(object, ...)

diagnose.gls(object, ...)
}
\arguments{
\item{object}{\link[nlme]{lme} or \link[nlme]{gls} object}

\item{...}{..}
}
\description{
..
}
\examples{
# lme or gls model
diagnose.lme(ortho1, lty = 2, id = .05, adj = -.4)
diagnose.lme(ortho2)
diagnose.lme(oxide1)
# diagnose.gls(ovary1) # bug

}
