% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_more.R
\name{model_frame_nlme}
\alias{model_frame_nlme}
\alias{model.frame.lme}
\alias{model.frame.gls}
\title{\link[stats]{model.frame} of \link[nlme]{gls} and \link[nlme]{lme} Objects}
\usage{
\method{model.frame}{lme}(formula, data = getData(formula), ...)

\method{model.frame}{gls}(formula, data = getData(formula), ...)
}
\arguments{
\item{formula}{\link[nlme]{gls} or \link[nlme]{lme} object}

\item{data}{..}

\item{...}{additional parameters of \link[stats]{model.frame.default}}
}
\description{
..
}
\details{
If not define, dispatch to \link[stats]{model.frame.default} and
get \verb{$modelStruct} (no \verb{$model} element).
}
